requirements.txt
# Deep Learning and ML
tensorflow==2.10.0
keras==2.10.0
torch==1.13.1
torchvision==0.14.1
scikit-learn==1.1.3

# Computer Vision
opencv-python==4.8.1.78
mediapipe==0.10.8
pillow==10.1.0

# Translation and TTS
googletrans==4.0.0rc1
gtts==2.4.0
TTS==0.20.0

# API and Web
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.5
python-socketio==5.10.0

# Data Processing
numpy==1.23.5
pandas==1.5.3
matplotlib==3.7.1
seaborn==0.12.2

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
pyautogui==0.9.54
pynput==1.7.6

# Jupyter
jupyter==1.0.0
ipykernel==6.25.0
notebook==7.0.6

# Testing
pytest==7.4.3
